Remove duplications in code, by combining the Bosses Animation Controller into one class.
